rename sidebar folder
