+++ src/pages/Marketplace/index.tsx
@@ -1,7 +1,7 @@
 import { useGetter } from "store/accessors";
 import Cards from "./Cards";
+import Filter from "./Filter";
 import Hero from "./Hero";
-import Sidebar from "./Sidebar";
 import Toolbar from "./Toolbar";
 
 export default function Marketplace() {
@@ -17,7 +17,7 @@ export default function Marketplace() {
 
       <div className="grid grid-cols-[auto_1fr] gap-x-8 grid-rows-[auto_1fr] padded-container min-h-screen">
         <Toolbar classes="my-10 col-span-2" />
-        <Sidebar
+        <Filter
           classes={`${
             isFilterOpen
               ? "fixed z-20 inset-0 md:grid md:relative md:z-0"
